# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Portfolio Platform API",
        "description": "Backend API for the AI-enabled dynamic portfolio platform",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:9100",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/admin/achievements/{id}": {
            "get": {
                "tags": [
                    "Achievements"
                ],
                "summary": "Get an achievement",
                "description": "Any status.",
                "operationId": "get_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAchievementResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Achievements"
                ],
                "summary": "Replace an achievement",
                "description": "Whole-row update.",
                "operationId": "update_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AchievementUpdateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAchievementResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Achievements"
                ],
                "summary": "Delete an achievement",
                "description": "Soft delete.",
                "operationId": "delete_6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/v1/analytics/page-view": {
            "post": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Record a page view",
                "description": "Anonymous and heavily rate-limited. Device and browser are derived server-side from the User-Agent; nothing identifying is stored.",
                "operationId": "record",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PageViewRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/achievements": {
            "get": {
                "tags": [
                    "Achievements"
                ],
                "summary": "List achievements",
                "description": "All statuses, optional ?status= filter.",
                "operationId": "list_7",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "PUBLISHED",
                                "ARCHIVED"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListAchievementResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Achievements"
                ],
                "summary": "Create an achievement",
                "description": "Defaults to PUBLISHED.",
                "operationId": "create_6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AchievementCreateRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseAchievementResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List published projects",
                "description": "Ordered by display order. Not paginated. Optional ?search= matches title, description and technology names (FR-18).",
                "operationId": "listPublished_1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListProjectSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/media/{id}/content": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Serve a file",
                "description": "Public: the Public Site renders these to anonymous visitors.",
                "operationId": "content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/admin/analytics/page-views": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Raw log",
                "description": "Newest first, filterable by entity type and date range.",
                "operationId": "pageViews",
                "parameters": [
                    {
                        "name": "entityType",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 50
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponsePageResponsePageViewResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MediaResponse": {
                "type": "object",
                "description": "An uploaded file",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "url": {
                        "type": "string",
                        "description": "API path that serves the bytes",
                        "example": "/api/v1/media/1/content"
                    },
                    "fileName": {
                        "type": "string"
                    },
                    "originalFileName": {
                        "type": "string"
                    },
                    "mimeType": {
                        "type": "string"
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "storageBackend": {
                        "type": "string",
                        "enum": [
                            "LOCAL",
                            "OBJECT_STORAGE"
                        ]
                    },
                    "width": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "height": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "altText": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "AchievementUpdateRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "achievedOn": {
                        "type": "string",
                        "format": "date"
                    },
                    "imageMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "aiVisible": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "title"
                ]
            },
            "AchievementResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "achievedOn": {
                        "type": "string",
                        "format": "date"
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "aiVisible": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseAchievementResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/AchievementResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {},
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "PageViewRequest": {
                "type": "object",
                "properties": {
                    "path": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    },
                    "entityType": {
                        "type": "string",
                        "description": "e.g. PROJECT",
                        "example": "PROJECT",
                        "maxLength": 30,
                        "minLength": 0
                    },
                    "entityId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "referrer": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 0
                    }
                },
                "required": [
                    "path"
                ]
            },
...
            "AchievementCreateRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 200,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string"
                    },
                    "achievedOn": {
                        "type": "string",
                        "format": "date"
                    },
                    "imageMediaId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    },
                    "aiVisible": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "title"
                ]
            },
...
            "ApiResponseListProjectSummaryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectSummaryResponse"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProjectSummaryResponse": {
                "type": "object",
                "description": "A project as it appears in a list",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "shortDescription": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "technologies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "projectType": {
                        "type": "string",
                        "enum": [
                            "PERSONAL",
                            "PROFESSIONAL",
                            "OPEN_SOURCE",
                            "CLIENT"
                        ]
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    }
                }
            },
...
            "ApiResponsePageResponsePageViewResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageResponsePageViewResponse"
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PageResponsePageViewResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PageViewResponse"
                        }
                    },
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "PageViewResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "path": {
                        "type": "string"
                    },
                    "entityType": {
                        "type": "string"
                    },
                    "entityId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "referrer": {
                        "type": "string"
                    },
                    "deviceType": {
                        "type": "string",
                        "enum": [
                            "DESKTOP",
                            "MOBILE",
                            "TABLET",
                            "UNKNOWN"
                        ]
                    },
                    "browser": {
                        "type": "string"
                    },
                    "viewedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ApiResponseListAchievementResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AchievementResponse"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    }
}